+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0
+
+Android integration complete:
+* send broadcasts on Android via Context.sendBroadcast
+
 ## 0.0.1
 
 Initial Release: